Provide parentQuestionId and required to PageLayout.tsx
In order to implement logic surrounding conditional display of question
and only allowing people to advance if they have completed all required questions,
these data fields have to be provided to the frontend. This commit
routes the data correctly.

To test it, I used console.log in PageLayout.tsx and confirmed the values
are both present and accurate. The parentQuestionId is present for the warehouse
questions and required questions had required set to true.